refactor: move all windows to own files
use event emitter to pass around info. disabled backdrops as they seem to be slowing down render of picker
