+++ src/renderers/cover/cover.html
@@ -4,9 +4,16 @@
 <head>
   <meta charset="utf-8" />
   <title>Overlay</title>
+
+  <style>
+    body {
+      background-color: rgba(0, 0, 0, 0.3);
+    }
+  </style>
 </head>
 
 <body>
+  <div id="background"></div>
   <script src="./cover.js"></script>
 </body>
